캠핑 상세 조회

+++ src/main/kotlin/kr/co/ats/camping/entity/camping/CampingInfo.kt
@@ -9,7 +9,7 @@ import javax.persistence.*
 @DynamicUpdate
 @Entity(name = "TB_CAMPING_INFO")
 class CampingInfo(
-    var avgRating:Long,
+    var avgRating:Double,
     @OneToOne(fetch = FetchType.EAGER, cascade = [CascadeType.ALL],mappedBy = "campingInfo")
     @JoinColumn(name = "CAMPING_INFO_KEY")
     var campingDetail: CampingDetail?,
@@ -19,7 +19,7 @@ class CampingInfo(
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "campingInfo")
     var campingReviewList:MutableList<CampingReview>?
     ) : BaseTimeEntity() {
-    constructor() : this(0L,null,null,null)
+    constructor() : this(0.0,null,null,null)
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
